Move validation check from `HalfEdge` to `Cycle`
I'm working on removing the reference to `Surface` from `HalfEdge`. Once
that is done, `HalfEdge` won't be able to support this validation check
anymore.
